Change various Int16 to Int32 because things need to be over 9000. Fixed knockback attribute name

+++ Item+CoreDataProperties.swift
@@ -16,21 +16,21 @@ extension Item {
         return NSFetchRequest<Item>(entityName: "Item");
     }
 
-    @NSManaged public var itemNo: Int16
+    @NSManaged public var itemNo: Int32
     @NSManaged public var itemName: String?
     @NSManaged public var flavorText: String?
     @NSManaged public var type: String?
     @NSManaged public var icon: NSData?
-    @NSManaged public var hp: Int16
-    @NSManaged public var ap: Int16
-    @NSManaged public var apr: Int16
-    @NSManaged public var atk: Int16
-    @NSManaged public var def: Int16
-    @NSManaged public var intl: Int16
-    @NSManaged public var res: Int16
-    @NSManaged public var hit: Int16
-    @NSManaged public var eva: Int16
-    @NSManaged public var mov: Int16
-    @NSManaged public var rng: Int16
+    @NSManaged public var hp: Int32
+    @NSManaged public var ap: Int32
+    @NSManaged public var apr: Int32
+    @NSManaged public var atk: Int32
+    @NSManaged public var def: Int32
+    @NSManaged public var intl: Int32
+    @NSManaged public var res: Int32
+    @NSManaged public var hit: Int32
+    @NSManaged public var eva: Int32
+    @NSManaged public var mov: Int32
+    @NSManaged public var rng: Int32
 
 }
